type fix and typo

+++ src/main/java/iguanaman/iguanatweakstconstruct/replacing/ReplacementLogic.java
@@ -140,7 +140,7 @@ public final class ReplacementLogic {
         updateTag(newTags, tags, "HarvestLevelHandle");
         updateTag(newTags, tags, "HarvestLevelExtra");
 
-        // ranged weapons have have additional tags to consider
+        // ranged weapons have additional tags to consider
         updateTag(newTags, tags, "DrawSpeed");
         updateTag(newTags, tags, "BaseDrawSpeed");
         updateTag(newTags, tags, "FlightSpeed");
@@ -224,7 +224,7 @@ public final class ReplacementLogic {
                 xp *= (100.0f - Config.partReplacementXpPenality) / 100.0f;
             }
 
-            tags.setInteger(LevelingLogic.TAG_EXP, Math.round(xp));
+            tags.setLong(LevelingLogic.TAG_EXP, Math.round(xp));
         }
 
         // handle boost leveling/xp
@@ -245,7 +245,7 @@ public final class ReplacementLogic {
                 xp *= (100.0f - Config.partReplacementBoostXpPenality) / 100.0f;
             }
 
-            tags.setInteger(LevelingLogic.TAG_BOOST_EXP, Math.round(xp));
+            tags.setLong(LevelingLogic.TAG_BOOST_EXP, Math.round(xp));
 
             // already full xp?
             if (LevelingLogic.isBoosted(tags)) tags.setInteger("HarvestLevel", tags.getInteger("HarvestLevel") + 1);
